refactor: 여러개의 컨트롤러를 만들어서, Command에 따른 실행 분리
