+++ src/main/java/chess/controller/SubController.java
@@ -1,0 +1,6 @@
+package chess.controller;
+
+public interface SubController {
+
+    void run();
+}
